retactor: 모임 찾기 구조 개선
