+++ src/components/features/composite/SearchFilterBar.tsx
@@ -39,7 +39,12 @@ export default function SearchFilterBar({
   };
 
   return (
-    <div className={cn("flex flex-col gap-4", className)}>
+    <div
+      className={cn(
+        "mb-6 flex flex-col gap-4 sm:mb-8 sm:gap-6 md:gap-8",
+        className,
+      )}
+    >
       <section className="flex flex-col gap-4 sm:flex-row sm:items-center sm:justify-between sm:gap-6">
         <div className="w-full sm:max-w-[480px]">
           <SearchBar
@@ -60,7 +65,7 @@ export default function SearchFilterBar({
 
       {currentKeyword && (
         <div className="text-sm font-medium text-slate-500">
-          <span className="text-main-purple font-bold">{`"${currentKeyword}"`}</span>{" "}
+          <span className="text-main-purple font-bold">{`"${currentKeyword}"`}</span>
           검색 결과입니다.
         </div>
       )}
